feat/reorganizacion\ close #35
