+++ backend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "opsie-tls-checker",
-  "module": "index.ts",
+  "module": "server.ts",
   "type": "module",
   "private": true,
   "devDependencies": {
